commentflow-ai/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic[email]
motor
google-auth-oauthlib
google-api-python-client
passlib
bcrypt
google-generativeai
